pdf2video/requirements.txt
pdf2image 
moviepy 
reportlab
pillow
PyMuPDF 
gTTS